use uint intead of uint64_t to prevent warning

+++ onnxruntime/python/onnxruntime_pybind_state.cc
@@ -1382,7 +1382,7 @@ including arg name, arg type (contains both type and shape).)pbdoc")
       .def("end_profiling", [](PyInferenceSession* sess) -> std::string {
         return sess->GetSessionHandle()->EndProfiling();
       })
-      .def_property_readonly("get_profiling_start_timeNs", [](const PyInferenceSession* sess) -> const uint64_t{
+      .def_property_readonly("get_profiling_start_timeNs", [](const PyInferenceSession* sess) -> const uint{
         return sess->GetSessionHandle()->GetProfiling().GetStartTimeNs();
       })
       .def("get_providers", [](PyInferenceSession* sess) -> const std::vector<std::string>& {
